Create metadata_redo.avdl
This is a new version of the metadata.vdl file. There are some rather extensive changes, some inline comments for consideration with indication of some general design decisions which may have to consider/rethink - especially if we want different records for ga4gh.Individual, ga4gh.Sample etc. or just one metadata object with a "Type" attribute. Also, if we keep the current structure a PR of this to the main branch should then have to include the notes for all attributes in all records.